cyber: protect the networked foothold from curriculum evolution
In a networked world (company / chain / any SSRF-pivot) the public SSRF
is the one mandatory foothold. soften offered "remove ssrf" and diversify
offered "swap ssrf -> ..." -- both strip the only entry, producing an
unsolvable world that admission rejects after the fact (wasted budget,
and it must never strip the entry). Completes PR1's evolution-safety pass.

Exclude the foothold SSRF from soften and diversify (source and swap-
target) when the world is networked; a single-service world, where ssrf
is an ordinary co-located vuln, is unaffected. Harden can still add an
SSRF decoy, and the world stays evolvable by every other move.

To use the existing networked check without a mutation<->__init__ import
cycle, _is_networked moves to sampling.py (re-exported from the package,
so every existing importer is unchanged) -- one definition, no duplicate.

Test: a networked world offers no ssrf soften/diversify yet stays
evolvable. Verified a minimal single-ssrf world isn't frozen. Full cyber
suite: 518 passed.
