+++ packs/cyber_webapp/cyber_webapp/sampling.py
@@ -275,6 +275,25 @@ _RECON_PATHS: tuple[str, ...] = (
     "/.well-known/app-config",
 )
 
+
+def _is_networked(graph: WorldGraph) -> bool:
+    # Networked = the flag is reachable only by pivoting: an SSRF on a PUBLIC service
+    # reaches an internal service that holds the flag. A vuln co-located with the flag
+    # on one service is not networked -- it stays single-container.
+    public_services = {
+        n.id for n in graph.by_kind("service") if n.attrs.get("exposure") == "public"
+    }
+    service_of_endpoint = {
+        e.dst: e.src for e in graph.edges.values() if e.kind == "exposes"
+    }
+    return any(
+        service_of_endpoint.get(edge.dst) in public_services
+        for vuln in graph.by_kind("vulnerability")
+        if vuln.attrs.get("kind") == "ssrf"
+        for edge in graph.out_edges(vuln.id, "affects")
+    )
+
+
 _COMMAND_INJECTION_BASE: tuple[str, ...] = (
     "ping",
     "nslookup",
